PythonProject/main.py
import os
import logging
import json
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup, LabeledPrice
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes, CallbackQueryHandler, PreCheckoutQueryHandler
from PIL import Image, ImageDraw, ImageFilter, ImageEnhance
import io
import asyncio
from datetime import datetime
from flask import Flask
import threading

# === НАСТРОЙКИ БОТА - ЗАМЕНИТЕ ЭТИ ЗНАЧЕНИЯ ===
BOT_TOKEN = "8346872542:AAGlnx0UaOZ0g6fmQDZLc9FL57fBeoiORTE"
BOT_OWNER_ID = 6523647184
PAYMENT_PROVIDER_TOKEN = "7490307358:TEST:CquDfTkRqNzVGqXS"
PROCESSING_COST = 35
# === КОНЕЦ НАСТРОЕК ===

# Создаем Flask приложение для поддержания активности
app = Flask(__name__)

@app.route('/')
def home():
    return "🤖 Bot is running!"

def run_flask():
    app.run(host='0.0.0.0', port=8080)

# Запускаем Flask в отдельном потоке
flask_thread = threading.Thread(target=run_flask)
flask_thread.daemon = True
flask_thread.start()

# Настройка логирования
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

# ... (ВСЯ ОСТАЛЬНАЯ ЧАСТЬ ВАШЕГО КОДА ОСТАЕТСЯ БЕЗ ИЗМЕНЕНИЙ)
# ВСЕ КЛАССЫ И ФУНКЦИИ ОСТАЮТСЯ ТАКИМИ ЖЕ
# PaymentSystem, ImageProcessor, TelegramBot - БЕЗ ИЗМЕНЕНИЙ

class PaymentSystem:
    # ... ваш существующий код без изменений
    pass

class ImageProcessor:
    # ... ваш существующий код без изменений
    pass

class TelegramBot:
    # ... ваш существующий код без изменений
    pass

# Запуск бота (ИЗМЕНЕННЫЙ БЛОК)
if __name__ == "__main__":
    print("✅ Все настройки корректны, запускаю бота...")
    print("🌐 Flask сервер запущен на порту 8080")
    try:
        bot = TelegramBot()
        bot.run()
    except Exception as e:
        print(f"❌ Ошибка запуска: {e}")